+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.27.0
 
 - Added `cloudName` to RelationshipSpecification
+- Updated `relationships` on `ServiceResource.relationships` to take `cloud_name` from the `RelationshipSpecification`
 
 
 # 0.27.2